webapi/tct-touchevent-w3c-tests/touchevent/Touch_attributes_exist.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xin,Liu <xinx.liu@intel.com>
        Wang, Hongjuan <hongjuanx.wang@intel.com>

-->

<html>
  <head>
    <title>TouchEvent test: Touch_attributes_exist</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2013/WD-touch-events-20130124/#idl-def-Touch" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body >
    <div id='touchdiv'>
    </div>
    <div id="log"></div>
    <script>
        var target = document.getElementById("touchdiv");
        var touch = document.createTouch(window, target, 1, 100, 101, 102, 103, 5, 3, 10, 10);

        test(function() {
            assert_true("identifier" in touch, "identifier attribute in Touch element");
        }, "Check if the Touch.identifier attribute exists");

        test(function() {
            assert_true("target" in touch, "target attribute in Touch element");
        }, "Check if the Touch.target attribute exists");

        test(function() {
            assert_true("screenX" in touch, "screenX attribute in Touch element");
        }, "Check if the Touch.screenX attribute exists");

        test(function() {
            assert_true("screenY" in touch, "screenY attribute in Touch element");
        }, "Check if the Touch.screenY attribute exists");

        test(function() {
            assert_true("clientX" in touch, "clientX attribute in Touch element");
        }, "Check if the Touch.clientX attribute exists");

        test(function() {
            assert_true("clientY" in touch, "clientY attribute in Touch element");
        }, "Check if the Touch.clientY attribute exists");

        test(function() {
            assert_true("pageX" in touch, "pageX attribute in Touch element");
        }, "Check if the Touch.pageX attribute exists");

        test(function() {
            assert_true("pageY" in touch, "pageY attribute in Touch element");
        }, "Check if the Touch.pageY attribute exists");
    </script>
  </body>
</html>
